requirements.txt
flask
flask_sqlalchemy
pyodbc
python-dotenv
pycryptodome